Add database (Postgres) to Express Compose
