+++ README.md
@@ -5,4 +5,4 @@ Compose Templates for me
 - [postgres](https://github.com/SongCastle/compose-templates/tree/main/postgres)
 - [mysql](https://github.com/SongCastle/compose-templates/tree/main/mysql)
 - [phoenix](https://github.com/SongCastle/compose-templates/tree/main/phoenix)
-- [express](https://github.com/SongCastle/compose-templates/tree/main/express)
+- [express-postgres](https://github.com/SongCastle/compose-templates/tree/main/express-postgres)
