+++ express-postgres/.env.sample
@@ -1,0 +1,13 @@
+# postgres
+POSTGRES_USER=user
+POSTGRES_PASSWORD_FILE=/run/secrets/db_password
+POSTGRES_DB=dev
+POSTGRES_SCHEMA=shared
+POSTGRES_PORT=5432
+LANG=ja_JP.UTF-8
+
+# app
+NODE_USER=node
+EXPRESS_HOST=0.0.0.0
+EXPRESS_PORT=3000
+DB_HOST=postgres
