+++ express-postgres/.gitignore
@@ -1,0 +1,3 @@
+.env
+password.txt
+tmp/
