+++ express-postgres/password.txt.sample
@@ -1,0 +1,1 @@
+password
