+++ express-postgres/postgres/Dockerfile
@@ -1,0 +1,3 @@
+FROM postgres:alpine3.19
+
+COPY ./docker-entrypoint-initdb.d /docker-entrypoint-initdb.d
