+++ express-postgres/postgres/docker-entrypoint-initdb.d/schema.sql.template
@@ -1,0 +1,9 @@
+CREATE TABLE IF NOT EXISTS "users" (
+  "uuid" UUID NOT NULL DEFAULT gen_random_uuid(),
+  "name" VARCHAR(48) NOT NULL,
+  "created_at" TIMESTAMP NOT NULL,
+  PRIMARY KEY ("uuid")
+);
+COMMENT ON TABLE "users" IS 'user table';
+COMMENT ON COLUMN "users"."uuid" IS 'primary key';
+COMMENT ON COLUMN "users"."name" IS 'user name';
