--- express/.env.sample
@@ -1,3 +1,0 @@
-NODE_USER=node
-EXPRESS_HOST=0.0.0.0
-EXPRESS_PORT=3000
