--- express/README.md
@@ -1,15 +1,0 @@
-# Express
-
-Docker Compose for Express
-
-## Getting Started
-
-```sh
-$ cp .env.sample .env
-
-$ docker compose build
-$ docker compose run --rm express yarn install
-$ docker compose up
-```
-
-Then access to `localhost:3000`.
