--- express/compose.yaml
@@ -1,16 +1,0 @@
-services:
-  express:
-    build:
-      context: .
-      dockerfile: ./Dockerfile
-      args:
-        NODE_USER: "${NODE_USER}"
-    env_file: ./.env
-    ports:
-      - 3000:${EXPRESS_PORT}
-    volumes:
-      - ./app:/usr/src/app
-      - no_modules:/usr/src/app/node_modules
-
-volumes:
-  no_modules:
